+++ src/activation_functions.rs
@@ -1,6 +1,6 @@
 use crate::fc_layer::*;
 
-pub fn activation_function(layers_cfg: &Vec<FCLayer>, i: usize, x: f64) -> f64 {
+pub fn activation_function(layers_cfg: &Vec<FCLayer>, i: usize, x: f32) -> f32 {
     /*let alert = match &layers_cfg[i].activation_function {
         sigmoid_function => println!("Applying a sigmoid function"),
         relu_function => println!("Applying a relu function"),
@@ -16,7 +16,7 @@ pub fn activation_function(layers_cfg: &Vec<FCLayer>, i: usize, x: f64) -> f64 {
     var
 }
 
-pub fn activation_function_prime(layers_cfg: &Vec<FCLayer>, i: usize, x: f64) -> f64 {
+pub fn activation_function_prime(layers_cfg: &Vec<FCLayer>, i: usize, x: f32) -> f32 {
     // let layer_activation_fn = layers_cfg[i].activation_function;
 
     let var = match &*layers_cfg[i].activation_function {
@@ -28,15 +28,15 @@ pub fn activation_function_prime(layers_cfg: &Vec<FCLayer>, i: usize, x: f64) ->
     var
 }
 
-pub fn sigmoid(x: f64) -> f64 {
+pub fn sigmoid(x: f32) -> f32 {
     1.0 / (1.0 + (-x).exp())
 }
 
-pub fn sigmoid_prime(x: f64) -> f64 {
+pub fn sigmoid_prime(x: f32) -> f32 {
     sigmoid(x) * (1.0 - sigmoid(x))
 }
 
-pub fn relu(x: f64) -> f64 {
+pub fn relu(x: f32) -> f32 {
     if x < 0.0 {
         0.0
     } else {
@@ -44,7 +44,7 @@ pub fn relu(x: f64) -> f64 {
     }
 }
 
-pub fn relu_prime(x: f64) -> f64 {
+pub fn relu_prime(x: f32) -> f32 {
     if x < 0.0 {
         0.0
     } else {
@@ -52,7 +52,7 @@ pub fn relu_prime(x: f64) -> f64 {
     }
 }
 
-pub fn threshold(x: f64, threshold: f64) -> f64 {
+pub fn threshold(x: f32, threshold: f32) -> f32 {
     if x > threshold {
         1.0
     } else {
